+++ pytorch_lightning/utilities/distributed.py
@@ -46,6 +46,7 @@ rank_zero_only.rank = getattr(rank_zero_only, 'rank', int(os.environ.get('LOCAL_
 
 
 def _warn(*args, **kwargs):
+
     warnings.warn(*args, **kwargs)